+++ apps/modernjs-ssr/nested-remote/CHANGELOG.md
@@ -1,5 +1,23 @@
 # modernjs-ssr-nested-remote
 
+## 0.1.20
+
+### Patch Changes
+
+- @module-federation/modern-js@0.6.0
+
+## 0.1.19
+
+### Patch Changes
+
+- @module-federation/modern-js@0.5.2
+
+## 0.1.18
+
+### Patch Changes
+
+- @module-federation/modern-js@0.5.1
+
 ## 0.1.17
 
 ### Patch Changes