+++ apps/modernjs-ssr/remote-new-version/CHANGELOG.md
@@ -1,5 +1,23 @@
 # modernjs-ssr-remote-new-version
 
+## 0.1.22
+
+### Patch Changes
+
+- @module-federation/modern-js@0.6.0
+
+## 0.1.21
+
+### Patch Changes
+
+- @module-federation/modern-js@0.5.2
+
+## 0.1.20
+
+### Patch Changes
+
+- @module-federation/modern-js@0.5.1
+
 ## 0.1.19
 
 ### Patch Changes