+++ packages/runtime/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @module-federation/runtime
 
+## 0.6.0
+
+### Patch Changes
+
+- Updated dependencies [1d9bb77]
+  - @module-federation/sdk@0.6.0
+
+## 0.5.2
+
+### Patch Changes
+
+- 24ba96e: fix: handle circular init shared
+- b90fa7d: feat: add shareStrategy option
+- Updated dependencies [b90fa7d]
+  - @module-federation/sdk@0.5.2
+
+## 0.5.1
+
+### Patch Changes
+
+- @module-federation/sdk@0.5.1
+
 ## 0.5.0
 
 ### Minor Changes